requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
numpy>=1.21.0
tqdm>=4.61.2
nltk>=3.6.2
matplotlib>=3.4.2
pandas>=1.3.0
scikit-learn>=0.24.2